requirements.txt
crewai==0.28.8
crewai-tools==0.3.0

# LangChain ecosystem (stable with CrewAI)
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langsmith==0.1.17

# Ollama API client
ollama==0.5.3

# Google integrations
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1

# Env management
python-dotenv==1.0.0
